requirements.txt
python-dotenv
langchain
langchain-cohere==0.1.9
langchain-chroma
pypdf
chromadb
pandas
pdfplumber
tqdm